chore: refactored module name
